+++ admin/tabs/cluster/instances/services/ServiceModel.ts
@@ -152,7 +152,13 @@ export class ServiceModel extends BaseInstanceModel {
     }
 
     private processRawData(r: PlainObject) {
-        const provider = r.name && r.name.startsWith('hoistCore') ? 'Hoist' : 'App';
+        // For Grails <=6, plugin is prefix in name.
+        // For Grails >7, we provide class to determine provider
+        // TODO: simplify when Hoist v34+ required.
+        const provider =
+            r.name?.startsWith('hoistCore') || r.className?.startsWith('io.xh.hoist')
+                ? 'Hoist'
+                : 'App';
         const displayName = lowerFirst(r.name.replace('hoistCore', ''));
         return {provider, displayName, ...r};
     }